+++ src/components/asset-browser/types/polyhaven.ts
@@ -214,3 +214,18 @@ export interface AssetAuthorInfo {
 	email: string | null
 	donate: string | null
 }
+
+export function isPolyHavenFileInfo(file: unknown): file is FileInfo {
+	if (typeof file !== 'object' || file === null) {
+		return false
+	}
+
+	return (
+		'url' in file &&
+		typeof file.url === 'string' &&
+		'md5' in file &&
+		typeof file.md5 === 'string' &&
+		'size' in file &&
+		typeof file.size === 'number'
+	)
+}